INFRA 1046 Add client servlet and rename module
