+++ gateway-ha/src/main/java/com/lyft/data/gateway/ha/caching/CachingDatabaseConnection.java
@@ -1,7 +1,11 @@
 package com.lyft.data.gateway.ha.caching;
 
 public abstract class CachingDatabaseConnection {
-  public abstract void shutdown();
+  public abstract String get(String key);
+
+  public abstract void set(String key, String value);
 
   public abstract void testConnection();
+
+  public abstract void shutdown();
 }
