+++ gateway-ha/src/main/java/com/lyft/data/gateway/ha/config/ProxyBackendConfiguration.java
@@ -1,6 +1,7 @@
 package com.lyft.data.gateway.ha.config;
 
-import com.lyft.data.proxyserver.ProxyServerConfiguration;
+import com.lyft.data.server.config.GatewayServerConfiguration;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
@@ -8,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 @Data
 @RequiredArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class ProxyBackendConfiguration extends ProxyServerConfiguration {
+public class ProxyBackendConfiguration extends GatewayServerConfiguration {
   private boolean active = true;
   private String routingGroup = "adhoc";
 }
