+++ gateway-ha/src/main/java/com/lyft/data/gateway/ha/handler/QueryIdCachingServerHandler.java
@@ -6,19 +6,19 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.base.Strings;
+import com.google.common.cache.CacheLoader.InvalidCacheLoadException;
 import com.google.common.io.CharStreams;
+import com.lyft.data.gateway.ha.caching.CachingDatabaseManager;
 import com.lyft.data.gateway.ha.router.QueryHistoryManager;
 import com.lyft.data.gateway.ha.router.RoutingManager;
-import com.lyft.data.proxyserver.ProxyHandler;
-import com.lyft.data.proxyserver.wrapper.MultiReadHttpServletRequest;
+import com.lyft.data.server.handler.ServerHandler;
+import com.lyft.data.server.wrapper.MultiReadHttpServletRequest;
 
 import java.io.EOFException;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.util.HashMap;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -29,86 +29,27 @@ import org.eclipse.jetty.http.HttpStatus;
 import org.eclipse.jetty.util.Callback;
 
 @Slf4j
-public class QueryIdCachingProxyHandler extends ProxyHandler {
-  public static final String PROXY_TARGET_HEADER = "proxytarget";
-  public static final String V1_STATEMENT_PATH = "/v1/statement";
-  public static final String V1_QUERY_PATH = "/v1/query";
-  public static final String V1_INFO_PATH = "/v1/info";
-  public static final String UI_API_STATS_PATH = "/ui/api/stats";
-  public static final String PRESTO_UI_PATH = "/ui";
-  // Add support for Trino
-  public static final String USER_HEADER = "X-Trino-User";
-  public static final String ALTERNATE_USER_HEADER = "X-Presto-User";
-  public static final String SOURCE_HEADER = "X-Trino-Source";
-  public static final String ALTERNATE_SOURCE_HEADER = "X-Presto-Source";
-  public static final String ROUTING_GROUP_HEADER = "X-Trino-Routing-Group";
-  public static final String ALTERNATE_ROUTING_GROUP_HEADER = "X-Presto-Routing-Group";
-  public static final String CLIENT_TAGS_HEADER = "X-Trino-Client-Tags";
-  public static final String ALTERNATE_CLIENT_TAGS_HEADER = "X-Presto-Client-Tags";
-  public static final String CONTENT_LENGTH_HEADER = "Content-Length";
-  public static final String ADHOC_ROUTING_GROUP = "adhoc";
-  private static final int QUERY_TEXT_LENGTH_FOR_HISTORY = 200;
-
-  private static final Pattern EXTRACT_BETWEEN_SINGLE_QUOTES = Pattern.compile("'([^\\s']+)'");
-
+public class QueryIdCachingServerHandler extends ServerHandler {
   private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
   private final RoutingManager routingManager;
   private final QueryHistoryManager queryHistoryManager;
+  private final CachingDatabaseManager cachingDatabaseManager;
 
   private final Meter requestMeter;
   private final int serverApplicationPort;
 
-  public QueryIdCachingProxyHandler(
+  public QueryIdCachingServerHandler(
       QueryHistoryManager queryHistoryManager,
       RoutingManager routingManager,
+      CachingDatabaseManager cachingDatabaseManager,
       int serverApplicationPort,
       Meter requestMeter) {
-    this.requestMeter = requestMeter;
-    this.routingManager = routingManager;
     this.queryHistoryManager = queryHistoryManager;
+    this.routingManager = routingManager;
+    this.cachingDatabaseManager = cachingDatabaseManager;
     this.serverApplicationPort = serverApplicationPort;
-  }
-
-  @Override
-  public void preConnectionHook(HttpServletRequest request, Request proxyRequest) {
-
-    log.debug("\n\nIN PRE CONN HOOK FOR " + request.getRequestURL() + "\n\n");
-    debugLogHeaders(proxyRequest);
-    debugLogHeaders(request);
-    System.out.println("\n\n");
-
-    if (request.getMethod().equals(HttpMethod.POST)
-        && request.getRequestURI().startsWith(V1_STATEMENT_PATH)) {
-      requestMeter.mark();
-      try {
-        String requestBody = CharStreams.toString(request.getReader());
-        // 6sense: Changed to debug to save sumo logic expenses
-        log.debug(
-            "Processing request endpoint: [{}], payload: [{}]",
-            request.getRequestURI(),
-            requestBody);
-        debugLogHeaders(request);
-      } catch (Exception e) {
-        log.warn("Error fetching the request payload", e);
-      }
-    }
-  }
-
-  private boolean isPathWhiteListed(String path) {
-    return path.startsWith(V1_STATEMENT_PATH)
-        || path.startsWith(V1_QUERY_PATH)
-        || path.startsWith(PRESTO_UI_PATH)
-        || path.startsWith(V1_INFO_PATH)
-        || path.startsWith(UI_API_STATS_PATH);
-  }
-
-  public boolean isAuthEnabled() {
-    return false;
-  }
-
-  public boolean handleAuthRequest(HttpServletRequest request) {
-    return true;
+    this.requestMeter = requestMeter;
   }
 
   @Override
@@ -170,56 +111,32 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
     return targetLocation;
   }
 
-  protected String extractQueryIdIfPresent(HttpServletRequest request) {
-    String path = request.getRequestURI();
-    String queryParams = request.getQueryString();
-    try {
-      String queryText = CharStreams.toString(request.getReader());
-      if (!Strings.isNullOrEmpty(queryText)
-          && queryText.toLowerCase().contains("system.runtime.kill_query")) {
-        // extract and return the queryId
-        String[] parts = queryText.split(",");
-        for (String part : parts) {
-          if (part.contains("query_id")) {
-            Matcher m = EXTRACT_BETWEEN_SINGLE_QUOTES.matcher(part);
-            if (m.find()) {
-              String queryQuoted = m.group();
-              if (!Strings.isNullOrEmpty(queryQuoted) && queryQuoted.length() > 0) {
-                return queryQuoted.substring(1, queryQuoted.length() - 1);
-              }
-            }
-          }
-        }
-      }
-    } catch (Exception e) {
-      log.error("Error extracting query payload from request", e);
-    }
-    if (path == null) {
-      return null;
-    }
-    String queryId = null;
+  @Override
+  public void preConnectionHook(HttpServletRequest request, Request proxyRequest) {
+    log.debug("\n\nIN PRE CONN HOOK FOR " + request.getRequestURL() + "\n\n");
+    debugLogHeaders(proxyRequest);
+    debugLogHeaders(request);
+    System.out.println("\n\n");
 
-    log.debug("trying to extract query id from path [{}] or queryString [{}]", path, queryParams);
-    if (path.startsWith(V1_STATEMENT_PATH) || path.startsWith(V1_QUERY_PATH)) {
-      String[] tokens = path.split("/");
-      if (tokens.length >= 4) {
-        if (path.contains("queued")
-            || path.contains("scheduled")
-            || path.contains("executing")
-            || path.contains("partialCancel")) {
-          queryId = tokens[4];
-        } else {
-          queryId = tokens[3];
-        }
+    if (request.getMethod().equals(HttpMethod.POST)
+        && request.getRequestURI().startsWith(V1_STATEMENT_PATH)) {
+      requestMeter.mark();
+      try {
+        String requestBody = CharStreams.toString(request.getReader());
+        // 6sense: Changed to debug to save sumo logic expenses
+        log.debug(
+            "Processing request endpoint: [{}], payload: [{}]",
+            request.getRequestURI(),
+            requestBody);
+        debugLogHeaders(request);
+      } catch (Exception e) {
+        log.warn("Error fetching the request payload", e);
       }
-    } else if (path.startsWith(PRESTO_UI_PATH)) {
-      queryId = queryParams;
     }
-    log.debug("query id in url [{}]", queryId);
-    return queryId;
   }
 
-  protected void postConnectionHook(
+  @Override
+  public void postConnectionHook(
       HttpServletRequest request,
       HttpServletResponse response,
       byte[] buffer,
@@ -239,17 +156,17 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
         boolean isGZipEncoding = isGZipEncoding(response);
         String output = "";
 
-        if (isGZipEncoding) {
-          output = plainTextFromGz(buffer);
-        } else {
-          output = new String(buffer);
-        }
-
-        log.debug("Response output [{}]", output);
-
         // This header is only contained on complete responses
         if (response.containsHeader(CONTENT_LENGTH_HEADER)) {
           try {
+            if (isGZipEncoding) {
+              output = plainTextFromGz(buffer);
+            } else {
+              output = new String(buffer);
+            }
+    
+            log.debug("Response output [{}]", output);
+
             JsonNode root = OBJECT_MAPPER.readTree(output);
   
             if (root.at("/error").isMissingNode()) {
@@ -284,8 +201,6 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
               log.debug("\n\nError Details:");
               log.debug(String.format("Error Code: %s ErrorName: %s Error Type: %s\n\n", 
                                 errorCode, errorName, errorType));
-  
-              // response.setStatus(503);
             }
           } catch (JsonParseException | EOFException e) {
             log.debug("Response does not contain complete information");
@@ -314,10 +229,18 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
             if (!Strings.isNullOrEmpty(queryDetail.getQueryId())) {
               routingManager.setBackendForQueryId(
                   queryDetail.getQueryId(), queryDetail.getBackendUrl());
+
               log.debug(
                   "QueryId [{}] mapped with proxy [{}]",
                   queryDetail.getQueryId(),
                   queryDetail.getBackendUrl());
+
+              // Caching response sent
+              cachingDatabaseManager.set(queryDetail.getQueryId()
+                    + CachingDatabaseManager.STALL_RESPONSE_SUFFIX, output);
+
+              // Saving history at gateway.
+              queryHistoryManager.submitQueryDetail(queryDetail);
             } else {
               log.debug("QueryId [{}] could not be cached", queryDetail.getQueryId());
             }
@@ -327,8 +250,6 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
                 output,
                 response.getStatus());
           }
-          // Saving history at gateway.
-          queryHistoryManager.submitQueryDetail(queryDetail);
         } else {
           log.debug("SKIPPING For {}", requestPath);
         }
@@ -340,7 +261,29 @@ public class QueryIdCachingProxyHandler extends ProxyHandler {
     super.postConnectionHook(request, response, buffer, offset, length, callback);
   }
 
-  private QueryHistoryManager.QueryDetail getQueryDetailsFromRequest(HttpServletRequest request)
+  @Override
+  public boolean fillResponseForQueryFromCache(HttpServletRequest req, HttpServletResponse resp) {
+    String queryId = extractQueryIdIfPresent(removeClientFromUri(req.getRequestURI()), req);
+
+    try {
+      String responseBody = cachingDatabaseManager.get(queryId 
+          + CachingDatabaseManager.STALL_RESPONSE_SUFFIX);
+
+      if (Strings.isNullOrEmpty(responseBody)) {
+        throw new InvalidCacheLoadException("No matching entry in cache");
+      }
+
+      fillRepsonseBody(responseBody.getBytes(Charset.defaultCharset()), resp);
+
+      return true;
+    } catch (Exception e) {
+      log.error("Error occured when returning response from cache for [{}]", queryId, e);
+    }
+
+    return false;
+  }
+
+  protected QueryHistoryManager.QueryDetail getQueryDetailsFromRequest(HttpServletRequest request)
       throws IOException {
     QueryHistoryManager.QueryDetail queryDetail = new QueryHistoryManager.QueryDetail();
     queryDetail.setBackendUrl(request.getHeader(PROXY_TARGET_HEADER));
