+++ server/pom.xml
@@ -12,7 +12,7 @@
         <relativePath>../</relativePath>
     </parent>
 
-    <artifactId>proxyserver</artifactId>
+    <artifactId>server</artifactId>
     <packaging>jar</packaging>
 
     <properties>
@@ -20,10 +20,16 @@
         <maven.compiler.target>1.8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <dropwizard.version>1.3.2</dropwizard.version>
     </properties>
 
 
     <dependencies>
+        <dependency>
+            <groupId>io.dropwizard</groupId>
+            <artifactId>dropwizard-core</artifactId>
+            <version>${dropwizard.version}</version>
+        </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
             <artifactId>jetty-server</artifactId>
